requirements.txt
streamlit
pandas
numpy
python-Levenshtein
python-dateutil
seaborn
matplotlib
networkx
pyvis
plotly
streamlit-aggrid
watchdog
phonetics
sentence-transformers
scikit-learn
umap-learn
scipy
datasketch
fuzzywuzzy